requirements.txt
billboard.py
flask
flask-sqlalchemy
python3-discogs-client